# 8308231: Faster MemAllocator::Allocation checks for verify/notification