@albertnetymk 2023-05-22
low-mem-detector and jfr are interested only in real-allocation (increasing mem usage), while dtrace will capture inside-tlab allocations also.

I think your change alters the behavior of dtrace, from capturing all-allocations to only jvmti-filtered ones. It can be surprising that the result of dtrace depends on some jvmti flags.

I reorganized the notify-methods and break them into three groups:

1. capturing all-allocations -- dtrace and jvmti-vm-obj-alloc
2. capturing real-allocation -- low-mem-detector and jfr
3. capturing real or pseudo alloc -- jvmti-sampler

The most important part looks like:

```cpp
  ~Allocation() {
    if (!check_out_of_memory()) {
      verify_after();

      notify_allocation_dtrace_sampler(_thread);

      // support for JVMTI VMObjectAlloc event (no-op if not enabled)
      JvmtiExport::vm_object_alloc_event_collector(obj());

      const bool is_real_allocation = _allocated_tlab_size != 0
                                   || _allocated_outside_tlab;

      if (is_real_allocation) {
        notify_allocation_low_memory_detector();
        notify_allocation_jfr_sampler();
      }

      if (is_real_allocation || _tlab_end_reset_for_sample) {
        notify_allocation_jvmti_sampler();
      }
    }
  }
```