@shipilev 2023-05-22
> I think your change alters the behavior of dtrace, from capturing all-allocations to only jvmti-filtered ones. It can be surprising that the result of dtrace depends on some jvmti flags.

Right. I just committed another variant, inspired by your suggestion. It passes testing for me, but I would be able to do performance tests only tomorrow.