@iklam 2023-08-11
> I force-pushed because there were lots of changes in the related code and the merge was exceedingly tedious. The `MultiArrayAlloc` benchmarks improve, I am going to post more thorough benchmark results next week. Meanwhile, @iklam, do you want to give this a spin with [JDK-8310823](https://bugs.openjdk.org/browse/JDK-8310823) prototype? I think it would be sensitive to this change as well.

I had one variant of my  [JDK-8310823](https://bugs.openjdk.org/browse/JDK-8310823) prototype that avoids MemAllocator altogether (calling `CollectedHeap::mem_allocate` directly). There was a very minor benefit. See https://github.com/openjdk/jdk/pull/14520#issuecomment-1650961800